do not keep v1 versions
