--- hydromt/data_adapter_v1/__init__.py
@@ -1,1 +1,0 @@
-"""Data Adapters are generic for its HydroMT type and perform transformations."""
